feat: Add base API namespace
