+++ resources/habit/__init__.py
@@ -1,0 +1,3 @@
+from flask_restx import Namespace
+
+habit_namespace = Namespace("habit", description="Habit API")
