+++ resources/routine/__init__.py
@@ -1,0 +1,3 @@
+from flask_restx import Namespace
+
+routine_namespace = Namespace("routine", description="Routine API")
