+++ resources/todo/__init__.py
@@ -1,0 +1,3 @@
+from flask_restx import Namespace
+
+todo_namespace = Namespace("todo", description="Todo API")
